refactor class and unit tests
