+++ .github/workflows/verify_library_generation.yaml
@@ -97,7 +97,7 @@ jobs:
     - name: Run python unit tests
       run: |
         set -x
-        python -m unittest library_generation/test/unit_tests.py
+        python -m unittest discover -s library_generation/test/ -p "*unit_tests.py"
   lint-shell:
     runs-on: ubuntu-22.04
     steps:
